refactor : Redis like, scrap, hits Service 분리 및 domain/redis Service 구분화
