+++ src/main/java/inu/codin/codin/domain/like/entity/LikeType.java
@@ -6,7 +6,8 @@ import lombok.Getter;
 public enum LikeType {
     POST("게시물"),
     COMMENT("댓글"),
-    REPLY("대댓글");
+    REPLY("대댓글"),
+    REVIEW("수강 후기");
 
     private final String description;
