+++ src/main/java/inu/codin/codin/infra/redis/config/RedisConfig.java
@@ -1,5 +1,6 @@
-package inu.codin.codin.infra.redis;
+package inu.codin.codin.infra.redis.config;
 
+import inu.codin.codin.infra.redis.dto.RedisProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.context.annotation.Bean;
